feat: Update Hibernate and fix major upgrade breaking changes
