+++ .gitignore
@@ -7,3 +7,8 @@ certs
 .project
 .settings
 bin/
+
+# Build artifacts
+gradle/
+gradlew
+gradlew.bat
