+++ src/main/resources/hibernate.cfg.xml
@@ -7,7 +7,6 @@
 
         <!-- Database connection settings -->
         <property name="hibernate.connection.driver_class">org.postgresql.Driver</property>
-        <property name="hibernate.dialect">org.hibernate.dialect.CockroachDB201Dialect</property>
         <property name="hibernate.connection.url">jdbc:postgresql://localhost:26257/defaultdb?sslmode=disable</property>
         <property name="hibernate.connection.username">root</property>
         <property name="hibernate.connection.password"></property>
@@ -19,5 +18,6 @@
         <!-- Optional: Show SQL output for debugging -->
         <property name="hibernate.show_sql">true</property>
         <property name="hibernate.format_sql">true</property>
+        <property name="hibernate.transaction.jta.platform">true</property>
     </session-factory>
 </hibernate-configuration>
